change data type of the filter

+++ project/data_analysis/python/get_spectra.py
@@ -57,9 +57,9 @@ for sv in surveys:
         shape, wcs = template[sv].data.shape, template[sv].data.wcs
         
         if ks_f["type"] == "binary_cross":
-            filter[sv] = so_map_preprocessing.build_std_filter(shape, wcs, vk_mask=ks_f["vk_mask"], hk_mask=ks_f["hk_mask"], dtype=np.float64)
+            filter[sv] = so_map_preprocessing.build_std_filter(shape, wcs, vk_mask=ks_f["vk_mask"], hk_mask=ks_f["hk_mask"], dtype=np.float32)
         elif ks_f["type"] == "gauss":
-            filter[sv] = so_map_preprocessing.build_sigurd_filter(shape, wcs, ks_f["lbounds"], dtype=np.float64)
+            filter[sv] = so_map_preprocessing.build_sigurd_filter(shape, wcs, ks_f["lbounds"], dtype=np.float32)
         else:
             print("you need to specify a valid filter type")
             sys.exit()
